characters in tile space now

+++ source/game.h
@@ -34,7 +34,7 @@ void Game_Destroy(game_t* game);
 void Game_Update(game_t* game, float delta_secs);
 void Game_Draw(game_t* game);
 
-character_t* Game_CreateCharacterAt(game_t* game, character_type_e type, bool is_player_controlled, u32 tile_x, u32 tile_y);
+character_t* Game_CreateCharacterAt(game_t* game, character_type_e type, bool is_player_controlled, float tile_x, float tile_y);
 
 void Game_ViewFocusTarget(game_t* game, vec2_t target);
 void Game_UpdateGameState(game_t* game, game_state_e state);
